+++ src/fields/text.cr
@@ -23,7 +23,7 @@ module Fields
     property is_unique : Bool
     # To optimize field traversal in the `paladins/check()` method.
     # Hint: It is recommended not to change.
-    getter group : UInt32 = 1
+    getter group : UInt8 = 1
 
     def initialize(
       @label : String = "",
